requirements.txt
fastapi
uvicorn[standard]
python-multipart
supabase
python-dotenv
openai
langchain
langchain-openai
langchain-pinecone
langchain-community
pinecone-client
tiktoken
email-validator
pydantic[email]
pypdf
